+++ pyflp/project.py
@@ -190,6 +190,8 @@ class Project(MultiEventModel):
     artists = EventProp[str](ProjectID.Artists)
     """Authors / artists info. to be embedded in exported WAV & MP3.
 
+    :menuselection:`Options --> &Project info --> Author`
+
     *New in FL Studio v5.0.*
     """
 
@@ -232,6 +234,8 @@ class Project(MultiEventModel):
     comments = EventProp[str](ProjectID.Comments, ProjectID._RTFComments)
     """Comments / project description / summary.
 
+    :menuselection:`Options --> &Project info --> Comments`
+
     Caution:
         Very old versions of FL used to store comments in RTF (Rich Text Format).
         PyFLP makes no efforts to parse that and stores it like a normal string
@@ -241,7 +245,10 @@ class Project(MultiEventModel):
     # Stored as a duration in days since the Delphi epoch (30 Dec, 1899).
     @property
     def created_on(self) -> datetime.datetime | None:
-        """The local date and time on which this project was created."""
+        """The local date and time on which this project was created.
+
+        Located at the bottom of :menuselection:`Options --> &Project info` page.
+        """
         if ProjectID.Timestamp in self._events:
             event = cast(TimestampEvent, self._events[ProjectID.Timestamp][0])
             return _DELPHI_EPOCH + datetime.timedelta(days=event["created_on"])
@@ -252,6 +259,8 @@ class Project(MultiEventModel):
     @property
     def data_path(self) -> pathlib.Path | None:
         """The absolute path used by FL to store all your renders.
+
+        :menuselection:`Options --> &Project general settings --> Data folder`
 
         *New in FL Studio v9.0.0.*
         """
@@ -272,6 +281,8 @@ class Project(MultiEventModel):
 
     genre = EventProp[str](ProjectID.Genre)
     """Genre of the song to be embedded in exported WAV & MP3.
+
+    :menuselection:`Options --> &Project info --> Genre`
 
     *New in FL Studio v5.0*.
     """
@@ -358,13 +369,18 @@ class Project(MultiEventModel):
         return Patterns(*self._collect_events(PatternsID, PatternID))
 
     pan_law = EventProp[PanLaw](ProjectID.PanLaw)
-    """Whether a circular or a triangular pan law is used for the project."""
+    """Whether a circular or a triangular pan law is used for the project.
+
+    :menuselection:`Options -> &Project general settings -> Advanced -> Panning law`
+    """
 
     @property
     def ppq(self) -> int:
         """Pulses per quarter.
 
         ![](https://bit.ly/3F0UrMT)
+
+        :menuselection:`Options --> &Project general settings --> Timebase (PPQ)`.
 
         Note:
             All types of lengths, positions and offsets internally use the PPQ
@@ -394,12 +410,17 @@ class Project(MultiEventModel):
     show_info = EventProp[bool](ProjectID.ShowInfo)
     """Whether to show a banner while the project is loading inside FL Studio.
 
+    :menuselection:`Options --> &Project info --> Show info on opening`
+
     The banner shows the :attr:`title`, :attr:`artists`, :attr:`genre`,
     :attr:`comments` and :attr:`url`.
     """
 
     title = EventProp[str](ProjectID.Title)
-    """Name of the song / project."""
+    """Name of the song / project.
+
+    :menuselection:`Options --> &Project info --> Title`
+    """
 
     # Stored internally as the actual BPM * 1000 as an integer.
     @property
@@ -464,12 +485,14 @@ class Project(MultiEventModel):
 
         ![](https://bit.ly/3TsBzdM)
 
+        Located at the bottom of :menuselection:`Options --> &Project info` page.
         """
         if ProjectID.Timestamp in self._events:
             event = cast(TimestampEvent, self._events[ProjectID.Timestamp][0])
             return datetime.timedelta(days=event["time_spent"])
 
     url = EventProp[str](ProjectID.Url)
+    """:menuselection:`Options --> &Project info --> Web link`."""
 
     # Internally represented as a string with a format of
     # `major.minor.patch.build?` *where `build` is optional, since older
@@ -483,6 +506,8 @@ class Project(MultiEventModel):
         """The version of FL Studio which was used to save the file.
 
         ![](https://bit.ly/3TD3BU0)
+
+        Located at the top of :menuselection:`Help --> &About` page.
 
         Caution:
             Changing this to a lower version will not make a file load magically